% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_error.R
\name{add_error}
\alias{add_error}
\title{Add error to sequence}
\usage{
add_error(dna, error)
}
\arguments{
\item{dna}{A character vector representing a DNA sequence, where each element
is a single DNA nucleotide.}

\item{error}{The desired error rate}
}
\value{
A DNA sequence string potentially containing errors at some
positions.
}
\description{
Simulate errors that happen in the DNA-based data storage process.
}
\examples{
DNAseq <- c("C", "G", "T", "G", "T", "A", "G", "A", "A", "T", "G", "T", "C")
add_error(DNAseq, 0.05)
}
