% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_to_dna.R
\name{bin_to_dna}
\alias{bin_to_dna}
\title{Transform binary code to DNA}
\usage{
bin_to_dna(bin, codevect)
}
\arguments{
\item{bin}{A character vector representing a binary sequence, where each
element contains a binary code corresponding to a single DNA nucleotide.}

\item{codevect}{A vector of length 4 with the correspondence
binary-to-quaternary code. The first element has the binary code
corresponding to A, the second to C, third to G, and fourth to T.}
}
\value{
A character vector representing a DNA sequence, where each element is
a single DNA nucleotide.
}
\description{
Transform binary code to DNA
}
\examples{
binseq <- c(1, 10, 11, 10, 11, 0, 10, 0, 0, 11, 10, 11, 1, 11, 11, 10, 10)
bin_to_dna(binseq, example_codevect)
}
