% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bin_to_quat.R
\name{bin_to_quat}
\alias{bin_to_quat}
\title{Get DNA sequence from binary string}
\usage{
bin_to_quat(binary, codevect)
}
\arguments{
\item{binary}{A binary code string (containing only '0's and '1's). It must
contain an even number of characters.}

\item{codevect}{A vector of length 4 with the correspondence
binary-to-quaternary code. The first element has the binary code
corresponding to A, the second to C, third to G, and fourth to T.}
}
\value{
A DNA-sequence string
}
\description{
Get DNA sequence from binary string
}
\examples{
binary <- "0110111011001000001110110111111010111001110000101100101100111011"
bin_to_quat(binary, example_codevect)

}
