% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dna_to_bin.R
\name{dna_to_bin}
\alias{dna_to_bin}
\title{Transform DNA code to binary}
\usage{
dna_to_bin(dna, codevect)
}
\arguments{
\item{dna}{A character vector representing a DNA sequence, where each element
is a single DNA nucleotide.}

\item{codevect}{A vector of length 4 with the correspondence
binary-to-quaternary code. The first element has the binary code
corresponding to A, the second to C, third to G, and fourth to T.}
}
\value{
A character vector representing a binary sequence, where each element
contains a binary code corresponding to a single DNA nucleotide.
}
\description{
Transform DNA code to binary
}
\examples{
DNAseq <- c("C", "G", "T", "G", "T", "A", "G", "A", "A", "T", "G", "T", "C")
v <- c("00", "01", "10", "11")
dna_to_bin(DNAseq, v)
}
