% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_bin}
\alias{example_bin}
\title{Example of binary encoding DNA}
\format{
A binary string containing 389 characters (0 or 1).
}
\source{
\url{https://doi.org/10.1049/iet-nbt.2020.0157}
}
\usage{
example_bin
}
\description{
A binary string that encodes a DNA sequence. In order to retrieve the DNA
sequence it contains, use the 'key' \code{example_codevect}:
\code{dna_to_quat(example_bin, example_codevect)}.
}
\keyword{datasets}
