% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_codevect}
\alias{example_codevect}
\title{Correspondence between binary and quaternary DNA code}
\format{
A vector of length 4 with binary strings as elements.
}
\source{
\url{https://doi.org/10.1049/iet-nbt.2020.0157}
}
\usage{
example_codevect
}
\description{
A vector of length 4 with the correspondence binary-to-quaternary code. The
first element has the binary code corresponding to A, the second to C, third
to G, and fourth to T. In this example, it is "00", "01", "10" and "11",
respectively.
}
\keyword{datasets}
