% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf_print.R
\name{pdf_print}
\alias{pdf_print}
\title{Split strings for PDF}
\usage{
pdf_print(x)
}
\arguments{
\item{x}{String to be split}
}
\description{
Split long strings so that they do not surpass margins in R markdown PDF
output.
}
