% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieval_XOR.R
\name{retrieval_XOR}
\alias{retrieval_XOR}
\title{Recover original DNA sequence from XOR fragments}
\usage{
retrieval_XOR(infile, codevect)
}
\arguments{
\item{infile}{The path of the FASTA file containing the sequences of the
three DNA fragments obtained by XOR encoding.}

\item{codevect}{A vector of length 4 with the correspondence
binary-to-quaternary code. The first element has the binary code
corresponding to A, the second to C, third to G, and fourth to T.}
}
\description{
This function simulates the retrieval step posterior to the DNA sequencing of
the three fragments created during encoding. The consensus method presented
here is not optimal for error correction; the best approach is not detailed
by Sharma et al.
}
