% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encoding_XOR.R
\name{encoding_XOR}
\alias{encoding_XOR}
\title{XOR encoding algorithm}
\usage{
encoding_XOR(binary, codevect, outfile)
}
\arguments{
\item{binary}{A binary code string (containing only '0's and '1's). It must
contain an even number of characters.}

\item{codevect}{A vector of length 4 with the correspondence
binary-to-quaternary code. The first element has the binary code
corresponding to A, the second to C, third to G, and fourth to T.}

\item{outfile}{The path where the FASTA file with the fragments will be saved.}
}
\description{
Simulating XOR encoding algorithm in DNA-based data storage process. The
resulting three DNA fragments are written to a FASTA file.
}
\examples{
encoding_XOR(example_bin, example_codevect, "../fragments.fa")
}
